feat(bikeshop): CRUD of brand (#1)
* feat(bikeshop): add brand entity

* feat(bikeshop): add migration to create the brands table

* feat(bikeshop): create a new brand

Create a DTO to validate the new brand
Create a service to save the new brand
Create a controller to accept a new store
Install class-validator and class-transformer

* feat(bikeshop): list all of the brands

* feat(bikeshop): get one brand by id

* fix(bikeshop): add a pipe to transform a numeric string to bigint

Add path alias to the common directory
Replace ParseIntPipe with ParseBigIntPipe

* fix(bikeshop): add an exception filter to catch EntityNotFoundError

Return a Not Found response

* feat(bikeshop): update one brand by id

Add DTO to validate the update changes
Install @nestjs/mapped-types

* feat(bikeshop): remove brand by id

* feat(bikeshop): add a pipe to transform a brand ID to a Brand

The pipe is responsible for transforming a numeric string representing a brand ID into a `Brand` object. It uses the `BrandService` to fetch the brand from the database based on the ID. If the brand is not found, it throws an `EntityNotFoundError`.